# src/management_console/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EndpointDLP",
    "description": "API for Endpoint Data Loss Prevention.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/agents/": {
      "get": {
        "tags": ["Agents"],
        "summary": "List Agents",
        "operationId": "list_agents_api_v1_agents__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/AgentStatus" },
                { "type": "null" }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Agents Api V1 Agents  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}": {
      "get": {
        "tags": ["Agents"],
        "summary": "Get Agent",
        "operationId": "get_agent_api_v1_agents__agent_id__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Agents"],
        "summary": "Update Agent",
        "operationId": "update_agent_api_v1_agents__agent_id__patch",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AgentCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Agents"],
        "summary": "Delete Agent",
        "operationId": "delete_agent_api_v1_agents__agent_id__delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/register": {
      "post": {
        "tags": ["Agents"],
        "summary": "Register Agent",
        "operationId": "register_agent_api_v1_agents_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AgentUpdate" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/agents/{agent_id}/heartbeat": {
      "patch": {
        "tags": ["Agents"],
        "summary": "Agent Heartbeat",
        "operationId": "agent_heartbeat_api_v1_agents__agent_id__heartbeat_patch",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          },
          {
            "name": "x-agent-key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "X-Agent-Key" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentCreate": {
        "properties": {
          "hostname": { "type": "string", "title": "Hostname" },
          "status": { "$ref": "#/components/schemas/AgentStatus" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["hostname", "status"],
        "title": "AgentCreate"
      },
...
      "AgentResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "hostname": { "type": "string", "title": "Hostname" },
          "status": { "$ref": "#/components/schemas/AgentStatus" },
          "last_seen": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Last Seen"
          },
          "updated_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Updated At"
          },
          "created_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Created At"
          },
          "policies": {
            "items": { "$ref": "#/components/schemas/PolicyResponse" },
            "type": "array",
            "title": "Policies",
            "default": []
          },
          "group_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Group Id"
          }
        },
        "type": "object",
        "required": ["id", "hostname", "status"],
        "title": "AgentResponse"
      },
      "AgentStatus": {
        "type": "string",
        "enum": ["active", "inactive", "offline"],
        "title": "AgentStatus"
      },
      "AgentUpdate": {
        "properties": {
          "hostname": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Hostname"
          },
          "status": {
            "anyOf": [
              { "$ref": "#/components/schemas/AgentStatus" },
              { "type": "null" }
            ]
          },
          "group_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Group Id"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "AgentUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyAction": {
        "type": "string",
        "enum": ["block", "alert", "allow"],
        "title": "PolicyAction"
      },
      "PolicyChanel": {
        "type": "string",
        "enum": ["all", "email", "clipboard", "browser", "usb"],
        "title": "PolicyChanel"
      },
...
      "PolicyResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "rule_type": { "$ref": "#/components/schemas/RuleType" },
          "rule": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rule"
          },
          "action": { "$ref": "#/components/schemas/PolicyAction" },
          "channel": { "$ref": "#/components/schemas/PolicyChanel" },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "rule_type",
          "rule",
          "action",
          "channel",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "PolicyResponse"
      },
...
      "RuleType": { "type": "string", "enum": ["regex"], "title": "RuleType" },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}